Refactored package names. Other naming convention violations will be handled later.
