+++ build.gradle
@@ -5,7 +5,7 @@ apply plugin: 'maven'
 apply plugin: 'java-library'
 
 //main class for running and variables
-mainClassName = 'Steg.StegMeister'
+mainClassName = 'steg.StegMeister'
 group 'StegMeister' //group name
 version = '1.0' //version number
 sourceCompatibility = 9 //java source version
@@ -107,7 +107,7 @@ jar {
     manifest {
         attributes 'Implementation-Title': 'StegMeister',
                    'Implementation-Version': version,
-                   'Main-Class': 'Steg.StegMeister'
+                   'Main-Class': 'steg.StegMeister'
     }
 }
