+++ docs/javadoc/Steg/Cipher.html
@@ -59,11 +59,11 @@
     <div class="subNav">
       <ul class="navList">
         <li>Prev&nbsp;Class</li>
-        <li><a href="../Steg/Controller.html" title="class in Steg"><span
+        <li><a href="../Steg/Controller.html" title="class in steg"><span
             class="typeNameLink">Next&nbsp;Class</span></a></li>
       </ul>
       <ul class="navList">
-        <li><a href="../index.html?Steg/Cipher.html" target="_top">Frames</a></li>
+        <li><a href="../index.html?steg/Cipher.html" target="_top">Frames</a></li>
         <li><a href="Cipher.html" target="_top">No&nbsp;Frames</a></li>
       </ul>
       <ul class="navList" id="allclasses_navbar_top">
@@ -117,7 +117,7 @@
   </script>
   <!-- ======== START OF CLASS DATA ======== -->
   <div class="header">
-    <div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../Steg/package-summary.html">Steg</a>
+    <div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../Steg/package-summary.html">steg</a>
     </div>
     <h2 title="Class Cipher" class="title">Class Cipher</h2>
   </div>
@@ -126,7 +126,7 @@
       <li>java.lang.Object</li>
       <li>
         <ul class="inheritance">
-          <li>Steg.Cryptography.Cryptography</li>
+          <li>steg.cryptography.Cryptography</li>
         </ul>
       </li>
     </ul>
@@ -135,7 +135,7 @@
         <li class="blockList">
           <dl>
             <dt>Direct Known Subclasses:</dt>
-            <dd><code><a href="../Steg/Decrypt.html" title="class in Steg">Decrypt</a></code></dd>
+            <dd><code><a href="../Steg/Decrypt.html" title="class in steg">Decrypt</a></code></dd>
           </dl>
           <hr>
           <br>
@@ -162,7 +162,7 @@ extends java.lang.Object</pre>
                   <th class="colLast" scope="col">Description</th>
                 </tr>
                 <tr class="altColor">
-                  <td class="colFirst"><code>(package private) <a href="../Steg/Encrypt.html" title="class in Steg">Encrypt</a></code>
+                  <td class="colFirst"><code>(package private) <a href="../Steg/Encrypt.html" title="class in steg">Encrypt</a></code>
                   </td>
                   <th class="colSecond" scope="row"><code><span class="memberNameLink"><a
                       href="../Steg/Cipher.html#encrypt">encrypt</a></span></code></th>
@@ -304,7 +304,7 @@ extends java.lang.Object</pre>
               <ul class="blockListLast">
                 <li class="blockList">
                   <h4>encrypt</h4>
-                  <pre><a href="../Steg/Encrypt.html" title="class in Steg">Encrypt</a> encrypt</pre>
+                  <pre><a href="../Steg/Encrypt.html" title="class in steg">Encrypt</a> encrypt</pre>
                 </li>
               </ul>
             </li>
@@ -417,11 +417,11 @@ extends java.lang.Object</pre>
   <div class="subNav">
     <ul class="navList">
       <li>Prev&nbsp;Class</li>
-      <li><a href="../Steg/Controller.html" title="class in Steg"><span class="typeNameLink">Next&nbsp;Class</span></a>
+      <li><a href="../Steg/Controller.html" title="class in steg"><span class="typeNameLink">Next&nbsp;Class</span></a>
       </li>
     </ul>
     <ul class="navList">
-      <li><a href="../index.html?Steg/Cipher.html" target="_top">Frames</a></li>
+      <li><a href="../index.html?steg/Cipher.html" target="_top">Frames</a></li>
       <li><a href="Cipher.html" target="_top">No&nbsp;Frames</a></li>
     </ul>
     <ul class="navList" id="allclasses_navbar_bottom">
