+++ docs/javadoc/Steg/Controller.html
@@ -58,13 +58,13 @@
     </div>
     <div class="subNav">
       <ul class="navList">
-        <li><a href="../Steg/Cipher.html" title="class in Steg"><span class="typeNameLink">Prev&nbsp;Class</span></a>
+        <li><a href="../Steg/Cipher.html" title="class in steg"><span class="typeNameLink">Prev&nbsp;Class</span></a>
         </li>
-        <li><a href="../Steg/Decrypt.html" title="class in Steg"><span class="typeNameLink">Next&nbsp;Class</span></a>
+        <li><a href="../Steg/Decrypt.html" title="class in steg"><span class="typeNameLink">Next&nbsp;Class</span></a>
         </li>
       </ul>
       <ul class="navList">
-        <li><a href="../index.html?Steg/Controller.html" target="_top">Frames</a></li>
+        <li><a href="../index.html?steg/Controller.html" target="_top">Frames</a></li>
         <li><a href="Controller.html" target="_top">No&nbsp;Frames</a></li>
       </ul>
       <ul class="navList" id="allclasses_navbar_top">
@@ -118,7 +118,7 @@
   </script>
   <!-- ======== START OF CLASS DATA ======== -->
   <div class="header">
-    <div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../Steg/package-summary.html">Steg</a>
+    <div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../Steg/package-summary.html">steg</a>
     </div>
     <h2 title="Class Controller" class="title">Class Controller</h2>
   </div>
@@ -127,7 +127,7 @@
       <li>java.lang.Object</li>
       <li>
         <ul class="inheritance">
-          <li>Steg.Interface.Controller</li>
+          <li>steg.ui.Controller</li>
         </ul>
       </li>
     </ul>
@@ -382,12 +382,12 @@ extends java.lang.Object</pre>
   </div>
   <div class="subNav">
     <ul class="navList">
-      <li><a href="../Steg/Cipher.html" title="class in Steg"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
-      <li><a href="../Steg/Decrypt.html" title="class in Steg"><span class="typeNameLink">Next&nbsp;Class</span></a>
+      <li><a href="../Steg/Cipher.html" title="class in steg"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+      <li><a href="../Steg/Decrypt.html" title="class in steg"><span class="typeNameLink">Next&nbsp;Class</span></a>
       </li>
     </ul>
     <ul class="navList">
-      <li><a href="../index.html?Steg/Controller.html" target="_top">Frames</a></li>
+      <li><a href="../index.html?steg/Controller.html" target="_top">Frames</a></li>
       <li><a href="Controller.html" target="_top">No&nbsp;Frames</a></li>
     </ul>
     <ul class="navList" id="allclasses_navbar_bottom">
