+++ docs/javadoc/Steg/Cryptography/Ciph.html
@@ -53,10 +53,10 @@ var pathtoroot = "../../";loadScripts(document, 'script');</script>
 <div class="subNav">
 <ul class="navList">
 <li>Prev&nbsp;Class</li>
-<li><a href="../../Steg/Cryptography/Cryptography.html" title="class in Steg.Cryptography"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+<li><a href="../../Steg/Cryptography/Cryptography.html" title="class in steg.cryptography"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Cryptography/Ciph.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Cryptography/Ciph.html" target="_top">Frames</a></li>
 <li><a href="Ciph.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_top">
@@ -110,7 +110,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </script>
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../../Steg/Cryptography/package-summary.html">Steg.Cryptography</a></div>
+<div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../../Steg/Cryptography/package-summary.html">steg.cryptography</a></div>
 <h2 title="Class Ciph" class="title">Class Ciph</h2>
 </div>
 <div class="contentContainer">
@@ -118,7 +118,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>Steg.Cryptography.Ciph</li>
+<li>steg.cryptography.Ciph</li>
 </ul>
 </li>
 </ul>
@@ -149,12 +149,12 @@ extends java.lang.Object</pre>
 <th class="colLast" scope="col">Description</th>
 </tr>
 <tr class="altColor">
-<td class="colFirst"><code><a href="../../Steg/Cryptography/Decrypt.html" title="class in Steg.Cryptography">Decrypt</a></code></td>
+<td class="colFirst"><code><a href="../../Steg/Cryptography/Decrypt.html" title="class in steg.cryptography">Decrypt</a></code></td>
 <th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="../../Steg/Cryptography/Ciph.html#Decrypt">Decrypt</a></span></code></th>
 <td class="colLast">&nbsp;</td>
 </tr>
 <tr class="rowColor">
-<td class="colFirst"><code><a href="../../Steg/Cryptography/Encrypt.html" title="class in Steg.Cryptography">Encrypt</a></code></td>
+<td class="colFirst"><code><a href="../../Steg/Cryptography/Encrypt.html" title="class in steg.cryptography">Encrypt</a></code></td>
 <th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="../../Steg/Cryptography/Ciph.html#Encrypt">Encrypt</a></span></code></th>
 <td class="colLast">&nbsp;</td>
 </tr>
@@ -213,7 +213,7 @@ extends java.lang.Object</pre>
 <ul class="blockList">
 <li class="blockList">
 <h4>Encrypt</h4>
-<pre>public&nbsp;<a href="../../Steg/Cryptography/Encrypt.html" title="class in Steg.Cryptography">Encrypt</a> Encrypt</pre>
+<pre>public&nbsp;<a href="../../Steg/Cryptography/Encrypt.html" title="class in steg.cryptography">Encrypt</a> Encrypt</pre>
 </li>
 </ul>
 <a name="Decrypt">
@@ -222,7 +222,7 @@ extends java.lang.Object</pre>
 <ul class="blockListLast">
 <li class="blockList">
 <h4>Decrypt</h4>
-<pre>public&nbsp;<a href="../../Steg/Cryptography/Decrypt.html" title="class in Steg.Cryptography">Decrypt</a> Decrypt</pre>
+<pre>public&nbsp;<a href="../../Steg/Cryptography/Decrypt.html" title="class in steg.cryptography">Decrypt</a> Decrypt</pre>
 </li>
 </ul>
 </li>
@@ -277,10 +277,10 @@ extends java.lang.Object</pre>
 <div class="subNav">
 <ul class="navList">
 <li>Prev&nbsp;Class</li>
-<li><a href="../../Steg/Cryptography/Cryptography.html" title="class in Steg.Cryptography"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+<li><a href="../../Steg/Cryptography/Cryptography.html" title="class in steg.cryptography"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Cryptography/Ciph.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Cryptography/Ciph.html" target="_top">Frames</a></li>
 <li><a href="Ciph.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_bottom">
