+++ docs/javadoc/Steg/Cryptography/Cryptography.html
@@ -58,11 +58,11 @@ var pathtoroot = "../../";loadScripts(document, 'script');</script>
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../Steg/Cryptography/Ciph.html" title="class in Steg.Cryptography"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
-<li><a href="../../Steg/Cryptography/Decrypt.html" title="class in Steg.Cryptography"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+<li><a href="../../Steg/Cryptography/Ciph.html" title="class in steg.cryptography"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../Steg/Cryptography/Decrypt.html" title="class in steg.cryptography"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Cryptography/Cryptography.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Cryptography/Cryptography.html" target="_top">Frames</a></li>
 <li><a href="Cryptography.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_top">
@@ -116,7 +116,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </script>
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../../Steg/Cryptography/package-summary.html">Steg.Cryptography</a></div>
+<div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../../Steg/Cryptography/package-summary.html">steg.cryptography</a></div>
 <h2 title="Class Cryptography" class="title">Class Cryptography</h2>
 </div>
 <div class="contentContainer">
@@ -124,7 +124,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>Steg.Cryptography.Cryptography</li>
+<li>steg.cryptography.Cryptography</li>
 </ul>
 </li>
 </ul>
@@ -133,7 +133,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li class="blockList">
 <dl>
 <dt>Direct Known Subclasses:</dt>
-<dd><code><a href="../../Steg/Cryptography/Decrypt.html" title="class in Steg.Cryptography">Decrypt</a></code>, <code><a href="../../Steg/Cryptography/Encrypt.html" title="class in Steg.Cryptography">Encrypt</a></code></dd>
+<dd><code><a href="../../Steg/Cryptography/Decrypt.html" title="class in steg.cryptography">Decrypt</a></code>, <code><a href="../../Steg/Cryptography/Encrypt.html" title="class in steg.cryptography">Encrypt</a></code></dd>
 </dl>
 <hr>
 <br>
@@ -341,11 +341,11 @@ extends java.lang.Object</pre>
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../Steg/Cryptography/Ciph.html" title="class in Steg.Cryptography"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
-<li><a href="../../Steg/Cryptography/Decrypt.html" title="class in Steg.Cryptography"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+<li><a href="../../Steg/Cryptography/Ciph.html" title="class in steg.cryptography"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../Steg/Cryptography/Decrypt.html" title="class in steg.cryptography"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Cryptography/Cryptography.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Cryptography/Cryptography.html" target="_top">Frames</a></li>
 <li><a href="Cryptography.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_bottom">
