+++ docs/javadoc/Steg/Cryptography/Encrypt.html
@@ -58,11 +58,11 @@ var pathtoroot = "../../";loadScripts(document, 'script');</script>
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../Steg/Cryptography/Decrypt.html" title="class in Steg.Cryptography"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../Steg/Cryptography/Decrypt.html" title="class in steg.cryptography"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
 <li>Next&nbsp;Class</li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Cryptography/Encrypt.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Cryptography/Encrypt.html" target="_top">Frames</a></li>
 <li><a href="Encrypt.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_top">
@@ -116,7 +116,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </script>
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../../Steg/Cryptography/package-summary.html">Steg.Cryptography</a></div>
+<div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../../Steg/Cryptography/package-summary.html">steg.cryptography</a></div>
 <h2 title="Class Encrypt" class="title">Class Encrypt</h2>
 </div>
 <div class="contentContainer">
@@ -124,10 +124,10 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li><a href="../../Steg/Cryptography/Cryptography.html" title="class in Steg.Cryptography">Steg.Cryptography.Cryptography</a></li>
+<li><a href="../../Steg/Cryptography/Cryptography.html" title="class in steg.cryptography">steg.cryptography.Cryptography</a></li>
 <li>
 <ul class="inheritance">
-<li>Steg.Cryptography.Encrypt</li>
+<li>steg.cryptography.Encrypt</li>
 </ul>
 </li>
 </ul>
@@ -139,7 +139,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <hr>
 <br>
 <pre>public class <span class="typeNameLabel">Encrypt</span>
-extends <a href="../../Steg/Cryptography/Cryptography.html" title="class in Steg.Cryptography">Cryptography</a></pre>
+extends <a href="../../Steg/Cryptography/Cryptography.html" title="class in steg.cryptography">Cryptography</a></pre>
 <div class="block">The encrypt class is used for all things encryption, to include key generation.</div>
 </li>
 </ul>
@@ -206,7 +206,7 @@ extends <a href="../../Steg/Cryptography/Cryptography.html" title="class in Steg
 <li class="blockList"><a name="methods.inherited.from.class.Steg.Cryptography.Cryptography">
 <!--   -->
 </a>
-<h3>Methods inherited from class&nbsp;Steg.Cryptography.<a href="../../Steg/Cryptography/Cryptography.html" title="class in Steg.Cryptography">Cryptography</a></h3>
+<h3>Methods inherited from class&nbsp;steg.cryptography.<a href="../../Steg/Cryptography/Cryptography.html" title="class in steg.cryptography">Cryptography</a></h3>
 <code><a href="../../Steg/Cryptography/Cryptography.html#getKey--">getKey</a>, <a href="../../Steg/Cryptography/Cryptography.html#maxKeySize--">maxKeySize</a>, <a href="../../Steg/Cryptography/Cryptography.html#saveKey-java.lang.String-">saveKey</a>, <a href="../../Steg/Cryptography/Cryptography.html#setKey-javax.crypto.SecretKey-">setKey</a></code></li>
 </ul>
 <ul class="blockList">
@@ -319,11 +319,11 @@ extends <a href="../../Steg/Cryptography/Cryptography.html" title="class in Steg
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../Steg/Cryptography/Decrypt.html" title="class in Steg.Cryptography"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../Steg/Cryptography/Decrypt.html" title="class in steg.cryptography"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
 <li>Next&nbsp;Class</li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Cryptography/Encrypt.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Cryptography/Encrypt.html" target="_top">Frames</a></li>
 <li><a href="Encrypt.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_bottom">
