+++ docs/javadoc/Steg/Cryptography/package-frame.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (9.0.4) on Tue Mar 06 21:02:28 CST 2018 -->
-<title>Steg.Cryptography</title>
+<title>steg.cryptography</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="date" content="2018-03-06">
 <link rel="stylesheet" type="text/css" href="../../stylesheet.css" title="Style">
@@ -18,14 +18,14 @@
 <script type="text/javascript" src="../../jquery/jquery-ui.js"></script>
 </head>
 <body>
-<h1 class="bar"><a href="../../Steg/Cryptography/package-summary.html" target="classFrame">Steg.Cryptography</a></h1>
+<h1 class="bar"><a href="../../Steg/Cryptography/package-summary.html" target="classFrame">steg.cryptography</a></h1>
 <div class="indexContainer">
 <h2 title="Classes">Classes</h2>
 <ul title="Classes">
-<li><a href="Ciph.html" title="class in Steg.Cryptography" target="classFrame">Ciph</a></li>
-<li><a href="Cryptography.html" title="class in Steg.Cryptography" target="classFrame">Cryptography</a></li>
-<li><a href="Decrypt.html" title="class in Steg.Cryptography" target="classFrame">Decrypt</a></li>
-<li><a href="Encrypt.html" title="class in Steg.Cryptography" target="classFrame">Encrypt</a></li>
+<li><a href="Ciph.html" title="class in steg.cryptography" target="classFrame">Ciph</a></li>
+<li><a href="Cryptography.html" title="class in steg.cryptography" target="classFrame">Cryptography</a></li>
+<li><a href="Decrypt.html" title="class in steg.cryptography" target="classFrame">Decrypt</a></li>
+<li><a href="Encrypt.html" title="class in steg.cryptography" target="classFrame">Encrypt</a></li>
 </ul>
 </div>
 </body>
