+++ docs/javadoc/Steg/Cryptography/package-summary.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (9.0.4) on Tue Mar 06 21:02:28 CST 2018 -->
-<title>Steg.Cryptography</title>
+<title>steg.cryptography</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="date" content="2018-03-06">
 <link rel="stylesheet" type="text/css" href="../../stylesheet.css" title="Style">
@@ -56,7 +56,7 @@ var pathtoroot = "../../";loadScripts(document, 'script');</script>
 <li><a href="../../Steg/Interface/package-summary.html">Next&nbsp;Package</a></li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Cryptography/package-summary.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Cryptography/package-summary.html" target="_top">Frames</a></li>
 <li><a href="package-summary.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_top">
@@ -94,7 +94,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 //-->
 </script>
 <div class="header">
-<h1 title="Package" class="title">Package&nbsp;Steg.Cryptography</h1>
+<h1 title="Package" class="title">Package&nbsp;steg.cryptography</h1>
 </div>
 <div class="contentContainer">
 <ul class="blockList">
@@ -107,21 +107,21 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </tr>
 <tbody>
 <tr class="altColor">
-<th class="colFirst" scope="row"><a href="../../Steg/Cryptography/Ciph.html" title="class in Steg.Cryptography">Ciph</a></th>
+<th class="colFirst" scope="row"><a href="../../Steg/Cryptography/Ciph.html" title="class in steg.cryptography">Ciph</a></th>
 <td class="colLast">&nbsp;</td>
 </tr>
 <tr class="rowColor">
-<th class="colFirst" scope="row"><a href="../../Steg/Cryptography/Cryptography.html" title="class in Steg.Cryptography">Cryptography</a></th>
+<th class="colFirst" scope="row"><a href="../../Steg/Cryptography/Cryptography.html" title="class in steg.cryptography">Cryptography</a></th>
 <td class="colLast">
 <div class="block">The cipher class it the main class used for encrypting/decrypting a file</div>
 </td>
 </tr>
 <tr class="altColor">
-<th class="colFirst" scope="row"><a href="../../Steg/Cryptography/Decrypt.html" title="class in Steg.Cryptography">Decrypt</a></th>
+<th class="colFirst" scope="row"><a href="../../Steg/Cryptography/Decrypt.html" title="class in steg.cryptography">Decrypt</a></th>
 <td class="colLast">&nbsp;</td>
 </tr>
 <tr class="rowColor">
-<th class="colFirst" scope="row"><a href="../../Steg/Cryptography/Encrypt.html" title="class in Steg.Cryptography">Encrypt</a></th>
+<th class="colFirst" scope="row"><a href="../../Steg/Cryptography/Encrypt.html" title="class in steg.cryptography">Encrypt</a></th>
 <td class="colLast">
 <div class="block">The encrypt class is used for all things encryption, to include key generation.</div>
 </td>
@@ -155,7 +155,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li><a href="../../Steg/Interface/package-summary.html">Next&nbsp;Package</a></li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Cryptography/package-summary.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Cryptography/package-summary.html" target="_top">Frames</a></li>
 <li><a href="package-summary.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_bottom">
