+++ docs/javadoc/Steg/Cryptography/package-tree.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (9.0.4) on Tue Mar 06 21:02:28 CST 2018 -->
-<title>Steg.Cryptography Class Hierarchy</title>
+<title>steg.cryptography Class Hierarchy</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="date" content="2018-03-06">
 <link rel="stylesheet" type="text/css" href="../../stylesheet.css" title="Style">
@@ -56,7 +56,7 @@ var pathtoroot = "../../";loadScripts(document, 'script');</script>
 <li><a href="../../Steg/Interface/package-tree.html">Next</a></li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Cryptography/package-tree.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Cryptography/package-tree.html" target="_top">Frames</a></li>
 <li><a href="package-tree.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_top">
@@ -94,7 +94,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 //-->
 </script>
 <div class="header">
-<h1 class="title">Hierarchy For Package Steg.Cryptography</h1>
+<h1 class="title">Hierarchy For Package steg.cryptography</h1>
 <span class="packageHierarchyLabel">Package Hierarchies:</span>
 <ul class="horizontal">
 <li><a href="../../overview-tree.html">All Packages</a></li>
@@ -105,11 +105,11 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <ul>
 <li class="circle">java.lang.Object
 <ul>
-<li class="circle">Steg.Cryptography.<a href="../../Steg/Cryptography/Ciph.html" title="class in Steg.Cryptography"><span class="typeNameLink">Ciph</span></a></li>
-<li class="circle">Steg.Cryptography.<a href="../../Steg/Cryptography/Cryptography.html" title="class in Steg.Cryptography"><span class="typeNameLink">Cryptography</span></a>
+<li class="circle">steg.cryptography.<a href="../../Steg/Cryptography/Ciph.html" title="class in steg.cryptography"><span class="typeNameLink">Ciph</span></a></li>
+<li class="circle">steg.cryptography.<a href="../../Steg/Cryptography/Cryptography.html" title="class in steg.cryptography"><span class="typeNameLink">Cryptography</span></a>
 <ul>
-<li class="circle">Steg.Cryptography.<a href="../../Steg/Cryptography/Decrypt.html" title="class in Steg.Cryptography"><span class="typeNameLink">Decrypt</span></a></li>
-<li class="circle">Steg.Cryptography.<a href="../../Steg/Cryptography/Encrypt.html" title="class in Steg.Cryptography"><span class="typeNameLink">Encrypt</span></a></li>
+<li class="circle">steg.cryptography.<a href="../../Steg/Cryptography/Decrypt.html" title="class in steg.cryptography"><span class="typeNameLink">Decrypt</span></a></li>
+<li class="circle">steg.cryptography.<a href="../../Steg/Cryptography/Encrypt.html" title="class in steg.cryptography"><span class="typeNameLink">Encrypt</span></a></li>
 </ul>
 </li>
 </ul>
@@ -140,7 +140,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li><a href="../../Steg/Interface/package-tree.html">Next</a></li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Cryptography/package-tree.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Cryptography/package-tree.html" target="_top">Frames</a></li>
 <li><a href="package-tree.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_bottom">
