+++ docs/javadoc/Steg/Decrypt.html
@@ -52,13 +52,13 @@
     </div>
     <div class="subNav">
       <ul class="navList">
-        <li><a href="../Steg/Controller.html" title="class in Steg"><span
+        <li><a href="../Steg/Controller.html" title="class in steg"><span
             class="typeNameLink">Prev&nbsp;Class</span></a></li>
-        <li><a href="../Steg/Encrypt.html" title="class in Steg"><span class="typeNameLink">Next&nbsp;Class</span></a>
+        <li><a href="../Steg/Encrypt.html" title="class in steg"><span class="typeNameLink">Next&nbsp;Class</span></a>
         </li>
       </ul>
       <ul class="navList">
-        <li><a href="../index.html?Steg/Decrypt.html" target="_top">Frames</a></li>
+        <li><a href="../index.html?steg/Decrypt.html" target="_top">Frames</a></li>
         <li><a href="Decrypt.html" target="_top">No&nbsp;Frames</a></li>
       </ul>
       <ul class="navList" id="allclasses_navbar_top">
@@ -112,7 +112,7 @@
   </script>
   <!-- ======== START OF CLASS DATA ======== -->
   <div class="header">
-    <div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../Steg/package-summary.html">Steg</a>
+    <div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../Steg/package-summary.html">steg</a>
     </div>
     <h2 title="Class Decrypt" class="title">Class Decrypt</h2>
   </div>
@@ -121,10 +121,10 @@
       <li>java.lang.Object</li>
       <li>
         <ul class="inheritance">
-          <li><a href="../Steg/Cipher.html" title="class in Steg">Steg.Cryptography.Cryptography</a></li>
+          <li><a href="../Steg/Cipher.html" title="class in steg">steg.cryptography.Cryptography</a></li>
           <li>
             <ul class="inheritance">
-              <li>Steg.Cryptography.Decrypt</li>
+              <li>steg.cryptography.Decrypt</li>
             </ul>
           </li>
         </ul>
@@ -136,7 +136,7 @@
           <hr>
           <br>
           <pre>public class <span class="typeNameLabel">Decrypt</span>
-extends <a href="../Steg/Cipher.html" title="class in Steg">Cipher</a></pre>
+extends <a href="../Steg/Cipher.html" title="class in steg">Cipher</a></pre>
         </li>
       </ul>
     </div>
@@ -153,8 +153,8 @@ extends <a href="../Steg/Cipher.html" title="class in Steg">Cipher</a></pre>
                 <li class="blockList"><a name="fields.inherited.from.class.Steg.Cipher">
                   <!--   -->
                 </a>
-                  <h3>Fields inherited from class&nbsp;Steg.<a href="../Steg/Cipher.html"
-                                                               title="class in Steg">Cipher</a></h3>
+                  <h3>Fields inherited from class&nbsp;steg.<a href="../Steg/Cipher.html"
+                                                               title="class in steg">Cipher</a></h3>
                   <code><a href="../Steg/Cipher.html#encrypt">encrypt</a>, <a href="../Steg/Cipher.html#key">key</a>, <a
                       href="../Steg/Cipher.html#msg">msg</a></code></li>
               </ul>
@@ -191,8 +191,8 @@ extends <a href="../Steg/Cipher.html" title="class in Steg">Cipher</a></pre>
                 <li class="blockList"><a name="methods.inherited.from.class.Steg.Cipher">
                   <!--   -->
                 </a>
-                  <h3>Methods inherited from class&nbsp;Steg.<a href="../Steg/Cipher.html"
-                                                                title="class in Steg">Cipher</a></h3>
+                  <h3>Methods inherited from class&nbsp;steg.<a href="../Steg/Cipher.html"
+                                                                title="class in steg">Cipher</a></h3>
                   <code><a href="../Steg/Cipher.html#getKey--">getKey</a>, <a
                       href="../Steg/Cipher.html#saveKey-java.lang.String-">saveKey</a>, <a
                       href="../Steg/Cipher.html#setKey-javax.crypto.SecretKey-">setKey</a>, <a
@@ -255,13 +255,13 @@ extends <a href="../Steg/Cipher.html" title="class in Steg">Cipher</a></pre>
   </div>
   <div class="subNav">
     <ul class="navList">
-      <li><a href="../Steg/Controller.html" title="class in Steg"><span class="typeNameLink">Prev&nbsp;Class</span></a>
+      <li><a href="../Steg/Controller.html" title="class in steg"><span class="typeNameLink">Prev&nbsp;Class</span></a>
       </li>
-      <li><a href="../Steg/Encrypt.html" title="class in Steg"><span class="typeNameLink">Next&nbsp;Class</span></a>
+      <li><a href="../Steg/Encrypt.html" title="class in steg"><span class="typeNameLink">Next&nbsp;Class</span></a>
       </li>
     </ul>
     <ul class="navList">
-      <li><a href="../index.html?Steg/Decrypt.html" target="_top">Frames</a></li>
+      <li><a href="../index.html?steg/Decrypt.html" target="_top">Frames</a></li>
       <li><a href="Decrypt.html" target="_top">No&nbsp;Frames</a></li>
     </ul>
     <ul class="navList" id="allclasses_navbar_bottom">
