+++ docs/javadoc/Steg/Encrypt.html
@@ -58,13 +58,13 @@
     </div>
     <div class="subNav">
       <ul class="navList">
-        <li><a href="../Steg/Decrypt.html" title="class in Steg"><span class="typeNameLink">Prev&nbsp;Class</span></a>
+        <li><a href="../Steg/Decrypt.html" title="class in steg"><span class="typeNameLink">Prev&nbsp;Class</span></a>
         </li>
-        <li><a href="../Steg/Steganography.html" title="class in Steg"><span class="typeNameLink">Next&nbsp;Class</span></a>
+        <li><a href="../Steg/Steganography.html" title="class in steg"><span class="typeNameLink">Next&nbsp;Class</span></a>
         </li>
       </ul>
       <ul class="navList">
-        <li><a href="../index.html?Steg/Encrypt.html" target="_top">Frames</a></li>
+        <li><a href="../index.html?steg/Encrypt.html" target="_top">Frames</a></li>
         <li><a href="Encrypt.html" target="_top">No&nbsp;Frames</a></li>
       </ul>
       <ul class="navList" id="allclasses_navbar_top">
@@ -118,7 +118,7 @@
   </script>
   <!-- ======== START OF CLASS DATA ======== -->
   <div class="header">
-    <div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../Steg/package-summary.html">Steg</a>
+    <div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../Steg/package-summary.html">steg</a>
     </div>
     <h2 title="Class Encrypt" class="title">Class Encrypt</h2>
   </div>
@@ -127,7 +127,7 @@
       <li>java.lang.Object</li>
       <li>
         <ul class="inheritance">
-          <li>Steg.Cryptography.Encrypt</li>
+          <li>steg.cryptography.Encrypt</li>
         </ul>
       </li>
     </ul>
@@ -294,13 +294,13 @@ extends java.lang.Object</pre>
   </div>
   <div class="subNav">
     <ul class="navList">
-      <li><a href="../Steg/Decrypt.html" title="class in Steg"><span class="typeNameLink">Prev&nbsp;Class</span></a>
+      <li><a href="../Steg/Decrypt.html" title="class in steg"><span class="typeNameLink">Prev&nbsp;Class</span></a>
       </li>
-      <li><a href="../Steg/Steganography.html" title="class in Steg"><span
+      <li><a href="../Steg/Steganography.html" title="class in steg"><span
           class="typeNameLink">Next&nbsp;Class</span></a></li>
     </ul>
     <ul class="navList">
-      <li><a href="../index.html?Steg/Encrypt.html" target="_top">Frames</a></li>
+      <li><a href="../index.html?steg/Encrypt.html" target="_top">Frames</a></li>
       <li><a href="Encrypt.html" target="_top">No&nbsp;Frames</a></li>
     </ul>
     <ul class="navList" id="allclasses_navbar_bottom">
