+++ docs/javadoc/Steg/Interface/Controller.html
@@ -62,7 +62,7 @@ var pathtoroot = "../../";loadScripts(document, 'script');</script>
 <li>Next&nbsp;Class</li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Interface/Controller.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Interface/Controller.html" target="_top">Frames</a></li>
 <li><a href="Controller.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_top">
@@ -116,7 +116,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </script>
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../../Steg/Interface/package-summary.html">Steg.Interface</a></div>
+<div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../../Steg/Interface/package-summary.html">steg.ui</a></div>
 <h2 title="Class Controller" class="title">Class Controller</h2>
 </div>
 <div class="contentContainer">
@@ -127,10 +127,10 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>javafx.application.Application</li>
 <li>
 <ul class="inheritance">
-<li><a href="../../Steg/StegMeister.html" title="class in Steg">Steg.StegMeister</a></li>
+<li><a href="../../Steg/StegMeister.html" title="class in steg">steg.StegMeister</a></li>
 <li>
 <ul class="inheritance">
-<li>Steg.Interface.Controller</li>
+<li>steg.ui.Controller</li>
 </ul>
 </li>
 </ul>
@@ -144,7 +144,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <hr>
 <br>
 <pre>public class <span class="typeNameLabel">Controller</span>
-extends <a href="../../Steg/StegMeister.html" title="class in Steg">StegMeister</a></pre>
+extends <a href="../../Steg/StegMeister.html" title="class in steg">StegMeister</a></pre>
 </li>
 </ul>
 </div>
@@ -254,7 +254,7 @@ extends <a href="../../Steg/StegMeister.html" title="class in Steg">StegMeister<
 <li class="blockList"><a name="methods.inherited.from.class.Steg.StegMeister">
 <!--   -->
 </a>
-<h3>Methods inherited from class&nbsp;Steg.<a href="../../Steg/StegMeister.html" title="class in Steg">StegMeister</a></h3>
+<h3>Methods inherited from class&nbsp;steg.<a href="../../Steg/StegMeister.html" title="class in steg">StegMeister</a></h3>
 <code><a href="../../Steg/StegMeister.html#getPrimaryStage--">getPrimaryStage</a>, <a href="../../Steg/StegMeister.html#main-java.lang.String:A-">main</a>, <a href="../../Steg/StegMeister.html#start-javafx.stage.Stage-">start</a></code></li>
 </ul>
 </li>
@@ -369,7 +369,7 @@ extends <a href="../../Steg/StegMeister.html" title="class in Steg">StegMeister<
 <li>Next&nbsp;Class</li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Interface/Controller.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Interface/Controller.html" target="_top">Frames</a></li>
 <li><a href="Controller.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_bottom">
