+++ docs/javadoc/Steg/Interface/package-summary.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (9.0.4) on Tue Mar 06 21:02:28 CST 2018 -->
-<title>Steg.Interface</title>
+<title>steg.ui</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="date" content="2018-03-06">
 <link rel="stylesheet" type="text/css" href="../../stylesheet.css" title="Style">
@@ -56,7 +56,7 @@ var pathtoroot = "../../";loadScripts(document, 'script');</script>
 <li><a href="../../Steg/Steganography/package-summary.html">Next&nbsp;Package</a></li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Interface/package-summary.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Interface/package-summary.html" target="_top">Frames</a></li>
 <li><a href="package-summary.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_top">
@@ -94,7 +94,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 //-->
 </script>
 <div class="header">
-<h1 title="Package" class="title">Package&nbsp;Steg.Interface</h1>
+<h1 title="Package" class="title">Package&nbsp;steg.ui</h1>
 </div>
 <div class="contentContainer">
 <ul class="blockList">
@@ -107,7 +107,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </tr>
 <tbody>
 <tr class="altColor">
-<th class="colFirst" scope="row"><a href="../../Steg/Interface/Controller.html" title="class in Steg.Interface">Controller</a></th>
+<th class="colFirst" scope="row"><a href="../../Steg/Interface/Controller.html" title="class in steg.ui">Controller</a></th>
 <td class="colLast">&nbsp;</td>
 </tr>
 </tbody>
@@ -139,7 +139,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li><a href="../../Steg/Steganography/package-summary.html">Next&nbsp;Package</a></li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Interface/package-summary.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Interface/package-summary.html" target="_top">Frames</a></li>
 <li><a href="package-summary.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_bottom">
