+++ docs/javadoc/Steg/Interface/package-tree.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (9.0.4) on Tue Mar 06 21:02:28 CST 2018 -->
-<title>Steg.Interface Class Hierarchy</title>
+<title>steg.ui Class Hierarchy</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="date" content="2018-03-06">
 <link rel="stylesheet" type="text/css" href="../../stylesheet.css" title="Style">
@@ -56,7 +56,7 @@ var pathtoroot = "../../";loadScripts(document, 'script');</script>
 <li><a href="../../Steg/Steganography/package-tree.html">Next</a></li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Interface/package-tree.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Interface/package-tree.html" target="_top">Frames</a></li>
 <li><a href="package-tree.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_top">
@@ -94,7 +94,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 //-->
 </script>
 <div class="header">
-<h1 class="title">Hierarchy For Package Steg.Interface</h1>
+<h1 class="title">Hierarchy For Package steg.ui</h1>
 <span class="packageHierarchyLabel">Package Hierarchies:</span>
 <ul class="horizontal">
 <li><a href="../../overview-tree.html">All Packages</a></li>
@@ -107,9 +107,9 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <ul>
 <li class="circle">javafx.application.Application
 <ul>
-<li class="circle">Steg.<a href="../../Steg/StegMeister.html" title="class in Steg"><span class="typeNameLink">StegMeister</span></a>
+<li class="circle">steg.<a href="../../Steg/StegMeister.html" title="class in steg"><span class="typeNameLink">StegMeister</span></a>
 <ul>
-<li class="circle">Steg.Interface.<a href="../../Steg/Interface/Controller.html" title="class in Steg.Interface"><span class="typeNameLink">Controller</span></a></li>
+<li class="circle">steg.ui.<a href="../../Steg/Interface/Controller.html" title="class in steg.ui"><span class="typeNameLink">Controller</span></a></li>
 </ul>
 </li>
 </ul>
@@ -142,7 +142,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li><a href="../../Steg/Steganography/package-tree.html">Next</a></li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Interface/package-tree.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Interface/package-tree.html" target="_top">Frames</a></li>
 <li><a href="package-tree.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_bottom">
