+++ docs/javadoc/Steg/StegMeister.html
@@ -62,7 +62,7 @@ var pathtoroot = "../";loadScripts(document, 'script');</script>
 <li>Next&nbsp;Class</li>
 </ul>
 <ul class="navList">
-<li><a href="../index.html?Steg/StegMeister.html" target="_top">Frames</a></li>
+<li><a href="../index.html?steg/StegMeister.html" target="_top">Frames</a></li>
 <li><a href="StegMeister.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_top">
@@ -116,7 +116,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </script>
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../Steg/package-summary.html">Steg</a></div>
+<div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../Steg/package-summary.html">steg</a></div>
 <h2 title="Class StegMeister" class="title">Class StegMeister</h2>
 </div>
 <div class="contentContainer">
@@ -127,7 +127,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>javafx.application.Application</li>
 <li>
 <ul class="inheritance">
-<li>Steg.StegMeister</li>
+<li>steg.StegMeister</li>
 </ul>
 </li>
 </ul>
@@ -138,7 +138,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li class="blockList">
 <dl>
 <dt>Direct Known Subclasses:</dt>
-<dd><code><a href="../Steg/Interface/Controller.html" title="class in Steg.Interface">Controller</a></code></dd>
+<dd><code><a href="../Steg/Interface/Controller.html" title="class in steg.ui">Controller</a></code></dd>
 </dl>
 <hr>
 <br>
@@ -338,7 +338,7 @@ extends javafx.application.Application</pre>
 <li>Next&nbsp;Class</li>
 </ul>
 <ul class="navList">
-<li><a href="../index.html?Steg/StegMeister.html" target="_top">Frames</a></li>
+<li><a href="../index.html?steg/StegMeister.html" target="_top">Frames</a></li>
 <li><a href="StegMeister.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_bottom">
