+++ docs/javadoc/Steg/Steganography.html
@@ -52,13 +52,13 @@
     </div>
     <div class="subNav">
       <ul class="navList">
-        <li><a href="../Steg/Encrypt.html" title="class in Steg"><span class="typeNameLink">Prev&nbsp;Class</span></a>
+        <li><a href="../Steg/Encrypt.html" title="class in steg"><span class="typeNameLink">Prev&nbsp;Class</span></a>
         </li>
-        <li><a href="../Steg/StegMeister.html" title="class in Steg"><span
+        <li><a href="../Steg/StegMeister.html" title="class in steg"><span
             class="typeNameLink">Next&nbsp;Class</span></a></li>
       </ul>
       <ul class="navList">
-        <li><a href="../index.html?Steg/Steganography.html" target="_top">Frames</a></li>
+        <li><a href="../index.html?steg/Steganography.html" target="_top">Frames</a></li>
         <li><a href="Steganography.html" target="_top">No&nbsp;Frames</a></li>
       </ul>
       <ul class="navList" id="allclasses_navbar_top">
@@ -112,7 +112,7 @@
   </script>
   <!-- ======== START OF CLASS DATA ======== -->
   <div class="header">
-    <div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../Steg/package-summary.html">Steg</a>
+    <div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../Steg/package-summary.html">steg</a>
     </div>
     <h2 title="Class Steganography" class="title">Class Steganography</h2>
   </div>
@@ -121,7 +121,7 @@
       <li>java.lang.Object</li>
       <li>
         <ul class="inheritance">
-          <li>Steg.Steganography</li>
+          <li>steg.steganography</li>
         </ul>
       </li>
     </ul>
@@ -222,13 +222,13 @@ extends java.lang.Object</pre>
   </div>
   <div class="subNav">
     <ul class="navList">
-      <li><a href="../Steg/Encrypt.html" title="class in Steg"><span class="typeNameLink">Prev&nbsp;Class</span></a>
+      <li><a href="../Steg/Encrypt.html" title="class in steg"><span class="typeNameLink">Prev&nbsp;Class</span></a>
       </li>
-      <li><a href="../Steg/StegMeister.html" title="class in Steg"><span class="typeNameLink">Next&nbsp;Class</span></a>
+      <li><a href="../Steg/StegMeister.html" title="class in steg"><span class="typeNameLink">Next&nbsp;Class</span></a>
       </li>
     </ul>
     <ul class="navList">
-      <li><a href="../index.html?Steg/Steganography.html" target="_top">Frames</a></li>
+      <li><a href="../index.html?steg/Steganography.html" target="_top">Frames</a></li>
       <li><a href="Steganography.html" target="_top">No&nbsp;Frames</a></li>
     </ul>
     <ul class="navList" id="allclasses_navbar_bottom">
