+++ docs/javadoc/Steg/Steganography/BasicDecoder.html
@@ -59,10 +59,10 @@ var pathtoroot = "../../";loadScripts(document, 'script');</script>
 <div class="subNav">
 <ul class="navList">
 <li>Prev&nbsp;Class</li>
-<li><a href="../../Steg/Steganography/BasicEncoder.html" title="class in Steg.Steganography"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+<li><a href="../../Steg/Steganography/BasicEncoder.html" title="class in steg.steganography"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Steganography/BasicDecoder.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Steganography/BasicDecoder.html" target="_top">Frames</a></li>
 <li><a href="BasicDecoder.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_top">
@@ -116,7 +116,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </script>
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../../Steg/Steganography/package-summary.html">Steg.Steganography</a></div>
+<div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../../Steg/Steganography/package-summary.html">steg.steganography</a></div>
 <h2 title="Class BasicDecoder" class="title">Class BasicDecoder</h2>
 </div>
 <div class="contentContainer">
@@ -124,7 +124,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>Steg.Steganography.BasicDecoder</li>
+<li>steg.steganography.BasicDecoder</li>
 </ul>
 </li>
 </ul>
@@ -254,10 +254,10 @@ extends java.lang.Object</pre>
 <div class="subNav">
 <ul class="navList">
 <li>Prev&nbsp;Class</li>
-<li><a href="../../Steg/Steganography/BasicEncoder.html" title="class in Steg.Steganography"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+<li><a href="../../Steg/Steganography/BasicEncoder.html" title="class in steg.steganography"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Steganography/BasicDecoder.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Steganography/BasicDecoder.html" target="_top">Frames</a></li>
 <li><a href="BasicDecoder.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_bottom">
