+++ docs/javadoc/Steg/Steganography/BasicEncoder.html
@@ -58,11 +58,11 @@ var pathtoroot = "../../";loadScripts(document, 'script');</script>
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../Steg/Steganography/BasicDecoder.html" title="class in Steg.Steganography"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
-<li><a href="../../Steg/Steganography/Model.html" title="class in Steg.Steganography"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+<li><a href="../../Steg/Steganography/BasicDecoder.html" title="class in steg.steganography"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../Steg/Steganography/Model.html" title="class in steg.steganography"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Steganography/BasicEncoder.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Steganography/BasicEncoder.html" target="_top">Frames</a></li>
 <li><a href="BasicEncoder.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_top">
@@ -116,7 +116,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </script>
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../../Steg/Steganography/package-summary.html">Steg.Steganography</a></div>
+<div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../../Steg/Steganography/package-summary.html">steg.steganography</a></div>
 <h2 title="Class BasicEncoder" class="title">Class BasicEncoder</h2>
 </div>
 <div class="contentContainer">
@@ -124,7 +124,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>Steg.Steganography.BasicEncoder</li>
+<li>steg.steganography.BasicEncoder</li>
 </ul>
 </li>
 </ul>
@@ -255,11 +255,11 @@ extends java.lang.Object</pre>
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../Steg/Steganography/BasicDecoder.html" title="class in Steg.Steganography"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
-<li><a href="../../Steg/Steganography/Model.html" title="class in Steg.Steganography"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+<li><a href="../../Steg/Steganography/BasicDecoder.html" title="class in steg.steganography"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../Steg/Steganography/Model.html" title="class in steg.steganography"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Steganography/BasicEncoder.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Steganography/BasicEncoder.html" target="_top">Frames</a></li>
 <li><a href="BasicEncoder.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_bottom">
