+++ docs/javadoc/Steg/Steganography/Model.html
@@ -52,11 +52,11 @@ var pathtoroot = "../../";loadScripts(document, 'script');</script>
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../Steg/Steganography/BasicEncoder.html" title="class in Steg.Steganography"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../Steg/Steganography/BasicEncoder.html" title="class in steg.steganography"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
 <li>Next&nbsp;Class</li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Steganography/Model.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Steganography/Model.html" target="_top">Frames</a></li>
 <li><a href="Model.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_top">
@@ -110,7 +110,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </script>
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../../Steg/Steganography/package-summary.html">Steg.Steganography</a></div>
+<div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../../Steg/Steganography/package-summary.html">steg.steganography</a></div>
 <h2 title="Class Model" class="title">Class Model</h2>
 </div>
 <div class="contentContainer">
@@ -118,7 +118,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>Steg.Steganography.Model</li>
+<li>steg.steganography.Model</li>
 </ul>
 </li>
 </ul>
@@ -149,12 +149,12 @@ extends java.lang.Object</pre>
 <th class="colLast" scope="col">Description</th>
 </tr>
 <tr class="altColor">
-<td class="colFirst"><code><a href="../../Steg/Steganography/BasicDecoder.html" title="class in Steg.Steganography">BasicDecoder</a></code></td>
+<td class="colFirst"><code><a href="../../Steg/Steganography/BasicDecoder.html" title="class in steg.steganography">BasicDecoder</a></code></td>
 <th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="../../Steg/Steganography/Model.html#decoder">decoder</a></span></code></th>
 <td class="colLast">&nbsp;</td>
 </tr>
 <tr class="rowColor">
-<td class="colFirst"><code><a href="../../Steg/Steganography/BasicEncoder.html" title="class in Steg.Steganography">BasicEncoder</a></code></td>
+<td class="colFirst"><code><a href="../../Steg/Steganography/BasicEncoder.html" title="class in steg.steganography">BasicEncoder</a></code></td>
 <th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="../../Steg/Steganography/Model.html#encoder">encoder</a></span></code></th>
 <td class="colLast">&nbsp;</td>
 </tr>
@@ -213,7 +213,7 @@ extends java.lang.Object</pre>
 <ul class="blockList">
 <li class="blockList">
 <h4>encoder</h4>
-<pre>public&nbsp;<a href="../../Steg/Steganography/BasicEncoder.html" title="class in Steg.Steganography">BasicEncoder</a> encoder</pre>
+<pre>public&nbsp;<a href="../../Steg/Steganography/BasicEncoder.html" title="class in steg.steganography">BasicEncoder</a> encoder</pre>
 </li>
 </ul>
 <a name="decoder">
@@ -222,7 +222,7 @@ extends java.lang.Object</pre>
 <ul class="blockListLast">
 <li class="blockList">
 <h4>decoder</h4>
-<pre>public&nbsp;<a href="../../Steg/Steganography/BasicDecoder.html" title="class in Steg.Steganography">BasicDecoder</a> decoder</pre>
+<pre>public&nbsp;<a href="../../Steg/Steganography/BasicDecoder.html" title="class in steg.steganography">BasicDecoder</a> decoder</pre>
 </li>
 </ul>
 </li>
@@ -269,11 +269,11 @@ extends java.lang.Object</pre>
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../Steg/Steganography/BasicEncoder.html" title="class in Steg.Steganography"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../Steg/Steganography/BasicEncoder.html" title="class in steg.steganography"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
 <li>Next&nbsp;Class</li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Steganography/Model.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Steganography/Model.html" target="_top">Frames</a></li>
 <li><a href="Model.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_bottom">
