+++ docs/javadoc/Steg/Steganography/package-frame.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (9.0.4) on Tue Mar 06 21:02:28 CST 2018 -->
-<title>Steg.Steganography</title>
+<title>steg.steganography</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="date" content="2018-03-06">
 <link rel="stylesheet" type="text/css" href="../../stylesheet.css" title="Style">
@@ -18,13 +18,13 @@
 <script type="text/javascript" src="../../jquery/jquery-ui.js"></script>
 </head>
 <body>
-<h1 class="bar"><a href="../../Steg/Steganography/package-summary.html" target="classFrame">Steg.Steganography</a></h1>
+<h1 class="bar"><a href="../../Steg/Steganography/package-summary.html" target="classFrame">steg.steganography</a></h1>
 <div class="indexContainer">
 <h2 title="Classes">Classes</h2>
 <ul title="Classes">
-<li><a href="BasicDecoder.html" title="class in Steg.Steganography" target="classFrame">BasicDecoder</a></li>
-<li><a href="BasicEncoder.html" title="class in Steg.Steganography" target="classFrame">BasicEncoder</a></li>
-<li><a href="Model.html" title="class in Steg.Steganography" target="classFrame">Model</a></li>
+<li><a href="BasicDecoder.html" title="class in steg.steganography" target="classFrame">BasicDecoder</a></li>
+<li><a href="BasicEncoder.html" title="class in steg.steganography" target="classFrame">BasicEncoder</a></li>
+<li><a href="Model.html" title="class in steg.steganography" target="classFrame">Model</a></li>
 </ul>
 </div>
 </body>
