+++ docs/javadoc/Steg/Steganography/package-summary.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (9.0.4) on Tue Mar 06 21:02:28 CST 2018 -->
-<title>Steg.Steganography</title>
+<title>steg.steganography</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="date" content="2018-03-06">
 <link rel="stylesheet" type="text/css" href="../../stylesheet.css" title="Style">
@@ -56,7 +56,7 @@ var pathtoroot = "../../";loadScripts(document, 'script');</script>
 <li>Next&nbsp;Package</li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Steganography/package-summary.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Steganography/package-summary.html" target="_top">Frames</a></li>
 <li><a href="package-summary.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_top">
@@ -94,7 +94,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 //-->
 </script>
 <div class="header">
-<h1 title="Package" class="title">Package&nbsp;Steg.Steganography</h1>
+<h1 title="Package" class="title">Package&nbsp;steg.steganography</h1>
 </div>
 <div class="contentContainer">
 <ul class="blockList">
@@ -107,15 +107,15 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </tr>
 <tbody>
 <tr class="altColor">
-<th class="colFirst" scope="row"><a href="../../Steg/Steganography/BasicDecoder.html" title="class in Steg.Steganography">BasicDecoder</a></th>
+<th class="colFirst" scope="row"><a href="../../Steg/Steganography/BasicDecoder.html" title="class in steg.steganography">BasicDecoder</a></th>
 <td class="colLast">&nbsp;</td>
 </tr>
 <tr class="rowColor">
-<th class="colFirst" scope="row"><a href="../../Steg/Steganography/BasicEncoder.html" title="class in Steg.Steganography">BasicEncoder</a></th>
+<th class="colFirst" scope="row"><a href="../../Steg/Steganography/BasicEncoder.html" title="class in steg.steganography">BasicEncoder</a></th>
 <td class="colLast">&nbsp;</td>
 </tr>
 <tr class="altColor">
-<th class="colFirst" scope="row"><a href="../../Steg/Steganography/Model.html" title="class in Steg.Steganography">Model</a></th>
+<th class="colFirst" scope="row"><a href="../../Steg/Steganography/Model.html" title="class in steg.steganography">Model</a></th>
 <td class="colLast">&nbsp;</td>
 </tr>
 </tbody>
@@ -147,7 +147,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>Next&nbsp;Package</li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Steganography/package-summary.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Steganography/package-summary.html" target="_top">Frames</a></li>
 <li><a href="package-summary.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_bottom">
