+++ docs/javadoc/Steg/Steganography/package-tree.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (9.0.4) on Tue Mar 06 21:02:28 CST 2018 -->
-<title>Steg.Steganography Class Hierarchy</title>
+<title>steg.steganography Class Hierarchy</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="date" content="2018-03-06">
 <link rel="stylesheet" type="text/css" href="../../stylesheet.css" title="Style">
@@ -56,7 +56,7 @@ var pathtoroot = "../../";loadScripts(document, 'script');</script>
 <li>Next</li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Steganography/package-tree.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Steganography/package-tree.html" target="_top">Frames</a></li>
 <li><a href="package-tree.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_top">
@@ -94,7 +94,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 //-->
 </script>
 <div class="header">
-<h1 class="title">Hierarchy For Package Steg.Steganography</h1>
+<h1 class="title">Hierarchy For Package steg.steganography</h1>
 <span class="packageHierarchyLabel">Package Hierarchies:</span>
 <ul class="horizontal">
 <li><a href="../../overview-tree.html">All Packages</a></li>
@@ -105,9 +105,9 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <ul>
 <li class="circle">java.lang.Object
 <ul>
-<li class="circle">Steg.Steganography.<a href="../../Steg/Steganography/BasicDecoder.html" title="class in Steg.Steganography"><span class="typeNameLink">BasicDecoder</span></a></li>
-<li class="circle">Steg.Steganography.<a href="../../Steg/Steganography/BasicEncoder.html" title="class in Steg.Steganography"><span class="typeNameLink">BasicEncoder</span></a></li>
-<li class="circle">Steg.Steganography.<a href="../../Steg/Steganography/Model.html" title="class in Steg.Steganography"><span class="typeNameLink">Model</span></a></li>
+<li class="circle">steg.steganography.<a href="../../Steg/Steganography/BasicDecoder.html" title="class in steg.steganography"><span class="typeNameLink">BasicDecoder</span></a></li>
+<li class="circle">steg.steganography.<a href="../../Steg/Steganography/BasicEncoder.html" title="class in steg.steganography"><span class="typeNameLink">BasicEncoder</span></a></li>
+<li class="circle">steg.steganography.<a href="../../Steg/Steganography/Model.html" title="class in steg.steganography"><span class="typeNameLink">Model</span></a></li>
 </ul>
 </li>
 </ul>
@@ -136,7 +136,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>Next</li>
 </ul>
 <ul class="navList">
-<li><a href="../../index.html?Steg/Steganography/package-tree.html" target="_top">Frames</a></li>
+<li><a href="../../index.html?steg/Steganography/package-tree.html" target="_top">Frames</a></li>
 <li><a href="package-tree.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_bottom">
