+++ docs/javadoc/Steg/package-frame.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (9.0.4) on Tue Mar 06 21:02:28 CST 2018 -->
-<title>Steg</title>
+<title>steg</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="date" content="2018-03-06">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
@@ -18,11 +18,11 @@
 <script type="text/javascript" src="../jquery/jquery-ui.js"></script>
 </head>
 <body>
-<h1 class="bar"><a href="../Steg/package-summary.html" target="classFrame">Steg</a></h1>
+<h1 class="bar"><a href="../Steg/package-summary.html" target="classFrame">steg</a></h1>
 <div class="indexContainer">
 <h2 title="Classes">Classes</h2>
 <ul title="Classes">
-<li><a href="StegMeister.html" title="class in Steg" target="classFrame">StegMeister</a></li>
+<li><a href="StegMeister.html" title="class in steg" target="classFrame">StegMeister</a></li>
 </ul>
 </div>
 </body>
