+++ docs/javadoc/allclasses-frame.html
@@ -21,15 +21,15 @@
 <h1 class="bar">All&nbsp;Classes</h1>
 <div class="indexContainer">
 <ul>
-<li><a href="Steg/Steganography/BasicDecoder.html" title="class in Steg.Steganography" target="classFrame">BasicDecoder</a></li>
-<li><a href="Steg/Steganography/BasicEncoder.html" title="class in Steg.Steganography" target="classFrame">BasicEncoder</a></li>
-<li><a href="Steg/Cryptography/Ciph.html" title="class in Steg.Cryptography" target="classFrame">Ciph</a></li>
-<li><a href="Steg/Interface/Controller.html" title="class in Steg.Interface" target="classFrame">Controller</a></li>
-<li><a href="Steg/Cryptography/Cryptography.html" title="class in Steg.Cryptography" target="classFrame">Cryptography</a></li>
-<li><a href="Steg/Cryptography/Decrypt.html" title="class in Steg.Cryptography" target="classFrame">Decrypt</a></li>
-<li><a href="Steg/Cryptography/Encrypt.html" title="class in Steg.Cryptography" target="classFrame">Encrypt</a></li>
-<li><a href="Steg/Steganography/Model.html" title="class in Steg.Steganography" target="classFrame">Model</a></li>
-<li><a href="Steg/StegMeister.html" title="class in Steg" target="classFrame">StegMeister</a></li>
+<li><a href="Steg/Steganography/BasicDecoder.html" title="class in steg.steganography" target="classFrame">BasicDecoder</a></li>
+<li><a href="Steg/Steganography/BasicEncoder.html" title="class in steg.steganography" target="classFrame">BasicEncoder</a></li>
+<li><a href="Steg/Cryptography/Ciph.html" title="class in steg.cryptography" target="classFrame">Ciph</a></li>
+<li><a href="Steg/Interface/Controller.html" title="class in steg.ui" target="classFrame">Controller</a></li>
+<li><a href="Steg/Cryptography/Cryptography.html" title="class in steg.cryptography" target="classFrame">Cryptography</a></li>
+<li><a href="Steg/Cryptography/Decrypt.html" title="class in steg.cryptography" target="classFrame">Decrypt</a></li>
+<li><a href="Steg/Cryptography/Encrypt.html" title="class in steg.cryptography" target="classFrame">Encrypt</a></li>
+<li><a href="Steg/Steganography/Model.html" title="class in steg.steganography" target="classFrame">Model</a></li>
+<li><a href="Steg/StegMeister.html" title="class in steg" target="classFrame">StegMeister</a></li>
 </ul>
 </div>
 </body>
