+++ docs/javadoc/allclasses-noframe.html
@@ -21,15 +21,15 @@
 <h1 class="bar">All&nbsp;Classes</h1>
 <div class="indexContainer">
 <ul>
-<li><a href="Steg/Steganography/BasicDecoder.html" title="class in Steg.Steganography">BasicDecoder</a></li>
-<li><a href="Steg/Steganography/BasicEncoder.html" title="class in Steg.Steganography">BasicEncoder</a></li>
-<li><a href="Steg/Cryptography/Ciph.html" title="class in Steg.Cryptography">Ciph</a></li>
-<li><a href="Steg/Interface/Controller.html" title="class in Steg.Interface">Controller</a></li>
-<li><a href="Steg/Cryptography/Cryptography.html" title="class in Steg.Cryptography">Cryptography</a></li>
-<li><a href="Steg/Cryptography/Decrypt.html" title="class in Steg.Cryptography">Decrypt</a></li>
-<li><a href="Steg/Cryptography/Encrypt.html" title="class in Steg.Cryptography">Encrypt</a></li>
-<li><a href="Steg/Steganography/Model.html" title="class in Steg.Steganography">Model</a></li>
-<li><a href="Steg/StegMeister.html" title="class in Steg">StegMeister</a></li>
+<li><a href="Steg/Steganography/BasicDecoder.html" title="class in steg.steganography">BasicDecoder</a></li>
+<li><a href="Steg/Steganography/BasicEncoder.html" title="class in steg.steganography">BasicEncoder</a></li>
+<li><a href="Steg/Cryptography/Ciph.html" title="class in steg.cryptography">Ciph</a></li>
+<li><a href="Steg/Interface/Controller.html" title="class in steg.ui">Controller</a></li>
+<li><a href="Steg/Cryptography/Cryptography.html" title="class in steg.cryptography">Cryptography</a></li>
+<li><a href="Steg/Cryptography/Decrypt.html" title="class in steg.cryptography">Decrypt</a></li>
+<li><a href="Steg/Cryptography/Encrypt.html" title="class in steg.cryptography">Encrypt</a></li>
+<li><a href="Steg/Steganography/Model.html" title="class in steg.steganography">Model</a></li>
+<li><a href="Steg/StegMeister.html" title="class in steg">StegMeister</a></li>
 </ul>
 </div>
 </body>
