+++ docs/javadoc/index-files/index-1.html
@@ -98,13 +98,13 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </a>
 <h2 class="title">B</h2>
 <dl>
-<dt><a href="../Steg/Steganography/BasicDecoder.html" title="class in Steg.Steganography"><span class="typeNameLink">BasicDecoder</span></a> - Class in <a href="../Steg/Steganography/package-summary.html">Steg.Steganography</a></dt>
+<dt><a href="../Steg/Steganography/BasicDecoder.html" title="class in steg.steganography"><span class="typeNameLink">BasicDecoder</span></a> - Class in <a href="../Steg/Steganography/package-summary.html">steg.steganography</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../Steg/Steganography/BasicDecoder.html#BasicDecoder--">BasicDecoder()</a></span> - Constructor for class Steg.Steganography.<a href="../Steg/Steganography/BasicDecoder.html" title="class in Steg.Steganography">BasicDecoder</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/Steganography/BasicDecoder.html#BasicDecoder--">BasicDecoder()</a></span> - Constructor for class steg.steganography.<a href="../Steg/Steganography/BasicDecoder.html" title="class in steg.steganography">BasicDecoder</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="../Steg/Steganography/BasicEncoder.html" title="class in Steg.Steganography"><span class="typeNameLink">BasicEncoder</span></a> - Class in <a href="../Steg/Steganography/package-summary.html">Steg.Steganography</a></dt>
+<dt><a href="../Steg/Steganography/BasicEncoder.html" title="class in steg.steganography"><span class="typeNameLink">BasicEncoder</span></a> - Class in <a href="../Steg/Steganography/package-summary.html">steg.steganography</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../Steg/Steganography/BasicEncoder.html#BasicEncoder--">BasicEncoder()</a></span> - Constructor for class Steg.Steganography.<a href="../Steg/Steganography/BasicEncoder.html" title="class in Steg.Steganography">BasicEncoder</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/Steganography/BasicEncoder.html#BasicEncoder--">BasicEncoder()</a></span> - Constructor for class steg.steganography.<a href="../Steg/Steganography/BasicEncoder.html" title="class in steg.steganography">BasicEncoder</a></dt>
 <dd>&nbsp;</dd>
 </dl>
 <a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">E</a>&nbsp;<a href="index-5.html">G</a>&nbsp;<a href="index-6.html">I</a>&nbsp;<a href="index-7.html">M</a>&nbsp;<a href="index-8.html">O</a>&nbsp;<a href="index-9.html">S</a>&nbsp;<a href="index-10.html">T</a>&nbsp;</div>
