+++ docs/javadoc/index-files/index-10.html
@@ -98,7 +98,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </a>
 <h2 class="title">T</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../Steg/Interface/Controller.html#test1btn--">test1btn()</a></span> - Method in class Steg.Interface.<a href="../Steg/Interface/Controller.html" title="class in Steg.Interface">Controller</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/Interface/Controller.html#test1btn--">test1btn()</a></span> - Method in class steg.ui.<a href="../Steg/Interface/Controller.html" title="class in steg.ui">Controller</a></dt>
 <dd>&nbsp;</dd>
 </dl>
 <a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">E</a>&nbsp;<a href="index-5.html">G</a>&nbsp;<a href="index-6.html">I</a>&nbsp;<a href="index-7.html">M</a>&nbsp;<a href="index-8.html">O</a>&nbsp;<a href="index-9.html">S</a>&nbsp;<a href="index-10.html">T</a>&nbsp;</div>
