+++ docs/javadoc/index-files/index-2.html
@@ -98,23 +98,23 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </a>
 <h2 class="title">C</h2>
 <dl>
-<dt><a href="../Steg/Cryptography/Ciph.html" title="class in Steg.Cryptography"><span class="typeNameLink">Ciph</span></a> - Class in <a href="../Steg/Cryptography/package-summary.html">Steg.Cryptography</a></dt>
+<dt><a href="../Steg/Cryptography/Ciph.html" title="class in steg.cryptography"><span class="typeNameLink">Ciph</span></a> - Class in <a href="../Steg/Cryptography/package-summary.html">steg.cryptography</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Ciph.html#Ciph--">Ciph()</a></span> - Constructor for class Steg.Cryptography.<a href="../Steg/Cryptography/Ciph.html" title="class in Steg.Cryptography">Ciph</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Ciph.html#Ciph--">Ciph()</a></span> - Constructor for class steg.cryptography.<a href="../Steg/Cryptography/Ciph.html" title="class in steg.cryptography">Ciph</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="../Steg/Interface/Controller.html" title="class in Steg.Interface"><span class="typeNameLink">Controller</span></a> - Class in <a href="../Steg/Interface/package-summary.html">Steg.Interface</a></dt>
+<dt><a href="../Steg/Interface/Controller.html" title="class in steg.ui"><span class="typeNameLink">Controller</span></a> - Class in <a href="../Steg/Interface/package-summary.html">steg.ui</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../Steg/Interface/Controller.html#Controller--">Controller()</a></span> - Constructor for class Steg.Interface.<a href="../Steg/Interface/Controller.html" title="class in Steg.Interface">Controller</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/Interface/Controller.html#Controller--">Controller()</a></span> - Constructor for class steg.ui.<a href="../Steg/Interface/Controller.html" title="class in steg.ui">Controller</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Encrypt.html#Crypt-java.lang.String-java.lang.String-">Crypt(String, String)</a></span> - Method in class Steg.Cryptography.<a href="../Steg/Cryptography/Encrypt.html" title="class in Steg.Cryptography">Encrypt</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Encrypt.html#Crypt-java.lang.String-java.lang.String-">Crypt(String, String)</a></span> - Method in class steg.cryptography.<a href="../Steg/Cryptography/Encrypt.html" title="class in steg.cryptography">Encrypt</a></dt>
 <dd>
 <div class="block">Default constructor to null</div>
 </dd>
-<dt><a href="../Steg/Cryptography/Cryptography.html" title="class in Steg.Cryptography"><span class="typeNameLink">Cryptography</span></a> - Class in <a href="../Steg/Cryptography/package-summary.html">Steg.Cryptography</a></dt>
+<dt><a href="../Steg/Cryptography/Cryptography.html" title="class in steg.cryptography"><span class="typeNameLink">Cryptography</span></a> - Class in <a href="../Steg/Cryptography/package-summary.html">steg.cryptography</a></dt>
 <dd>
 <div class="block">The cipher class it the main class used for encrypting/decrypting a file</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Cryptography.html#Cryptography--">Cryptography()</a></span> - Constructor for class Steg.Cryptography.<a href="../Steg/Cryptography/Cryptography.html" title="class in Steg.Cryptography">Cryptography</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Cryptography.html#Cryptography--">Cryptography()</a></span> - Constructor for class steg.cryptography.<a href="../Steg/Cryptography/Cryptography.html" title="class in steg.cryptography">Cryptography</a></dt>
 <dd>
 <div class="block">Default constructor to null</div>
 </dd>
