+++ docs/javadoc/index-files/index-3.html
@@ -98,15 +98,15 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </a>
 <h2 class="title">D</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../Steg/Steganography/BasicDecoder.html#decodeImage-javafx.scene.image.Image-">decodeImage(Image)</a></span> - Method in class Steg.Steganography.<a href="../Steg/Steganography/BasicDecoder.html" title="class in Steg.Steganography">BasicDecoder</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/Steganography/BasicDecoder.html#decodeImage-javafx.scene.image.Image-">decodeImage(Image)</a></span> - Method in class steg.steganography.<a href="../Steg/Steganography/BasicDecoder.html" title="class in steg.steganography">BasicDecoder</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../Steg/Steganography/Model.html#decoder">decoder</a></span> - Variable in class Steg.Steganography.<a href="../Steg/Steganography/Model.html" title="class in Steg.Steganography">Model</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/Steganography/Model.html#decoder">decoder</a></span> - Variable in class steg.steganography.<a href="../Steg/Steganography/Model.html" title="class in steg.steganography">Model</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="../Steg/Cryptography/Decrypt.html" title="class in Steg.Cryptography"><span class="typeNameLink">Decrypt</span></a> - Class in <a href="../Steg/Cryptography/package-summary.html">Steg.Cryptography</a></dt>
+<dt><a href="../Steg/Cryptography/Decrypt.html" title="class in steg.cryptography"><span class="typeNameLink">Decrypt</span></a> - Class in <a href="../Steg/Cryptography/package-summary.html">steg.cryptography</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Ciph.html#Decrypt">Decrypt</a></span> - Variable in class Steg.Cryptography.<a href="../Steg/Cryptography/Ciph.html" title="class in Steg.Cryptography">Ciph</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Ciph.html#Decrypt">Decrypt</a></span> - Variable in class steg.cryptography.<a href="../Steg/Cryptography/Ciph.html" title="class in steg.cryptography">Ciph</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Decrypt.html#Decrypt--">Decrypt()</a></span> - Constructor for class Steg.Cryptography.<a href="../Steg/Cryptography/Decrypt.html" title="class in Steg.Cryptography">Decrypt</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Decrypt.html#Decrypt--">Decrypt()</a></span> - Constructor for class steg.cryptography.<a href="../Steg/Cryptography/Decrypt.html" title="class in steg.cryptography">Decrypt</a></dt>
 <dd>
 <div class="block">Default constructor to null</div>
 </dd>
