+++ docs/javadoc/index-files/index-4.html
@@ -98,17 +98,17 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </a>
 <h2 class="title">E</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../Steg/Steganography/BasicEncoder.html#encodeImage-javafx.scene.image.Image-java.lang.String-">encodeImage(Image, String)</a></span> - Method in class Steg.Steganography.<a href="../Steg/Steganography/BasicEncoder.html" title="class in Steg.Steganography">BasicEncoder</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/Steganography/BasicEncoder.html#encodeImage-javafx.scene.image.Image-java.lang.String-">encodeImage(Image, String)</a></span> - Method in class steg.steganography.<a href="../Steg/Steganography/BasicEncoder.html" title="class in steg.steganography">BasicEncoder</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../Steg/Steganography/Model.html#encoder">encoder</a></span> - Variable in class Steg.Steganography.<a href="../Steg/Steganography/Model.html" title="class in Steg.Steganography">Model</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/Steganography/Model.html#encoder">encoder</a></span> - Variable in class steg.steganography.<a href="../Steg/Steganography/Model.html" title="class in steg.steganography">Model</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="../Steg/Cryptography/Encrypt.html" title="class in Steg.Cryptography"><span class="typeNameLink">Encrypt</span></a> - Class in <a href="../Steg/Cryptography/package-summary.html">Steg.Cryptography</a></dt>
+<dt><a href="../Steg/Cryptography/Encrypt.html" title="class in steg.cryptography"><span class="typeNameLink">Encrypt</span></a> - Class in <a href="../Steg/Cryptography/package-summary.html">steg.cryptography</a></dt>
 <dd>
 <div class="block">The encrypt class is used for all things encryption, to include key generation.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Ciph.html#Encrypt">Encrypt</a></span> - Variable in class Steg.Cryptography.<a href="../Steg/Cryptography/Ciph.html" title="class in Steg.Cryptography">Ciph</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Ciph.html#Encrypt">Encrypt</a></span> - Variable in class steg.cryptography.<a href="../Steg/Cryptography/Ciph.html" title="class in steg.cryptography">Ciph</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Encrypt.html#Encrypt--">Encrypt()</a></span> - Constructor for class Steg.Cryptography.<a href="../Steg/Cryptography/Encrypt.html" title="class in Steg.Cryptography">Encrypt</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Encrypt.html#Encrypt--">Encrypt()</a></span> - Constructor for class steg.cryptography.<a href="../Steg/Cryptography/Encrypt.html" title="class in steg.cryptography">Encrypt</a></dt>
 <dd>
 <div class="block">Default constructor to null</div>
 </dd>
