+++ docs/javadoc/index-files/index-5.html
@@ -98,15 +98,15 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </a>
 <h2 class="title">G</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Encrypt.html#genKey-int-">genKey(int)</a></span> - Method in class Steg.Cryptography.<a href="../Steg/Cryptography/Encrypt.html" title="class in Steg.Cryptography">Encrypt</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Encrypt.html#genKey-int-">genKey(int)</a></span> - Method in class steg.cryptography.<a href="../Steg/Cryptography/Encrypt.html" title="class in steg.cryptography">Encrypt</a></dt>
 <dd>
 <div class="block">Generate a random secret key</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Cryptography.html#getKey--">getKey()</a></span> - Method in class Steg.Cryptography.<a href="../Steg/Cryptography/Cryptography.html" title="class in Steg.Cryptography">Cryptography</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Cryptography.html#getKey--">getKey()</a></span> - Method in class steg.cryptography.<a href="../Steg/Cryptography/Cryptography.html" title="class in steg.cryptography">Cryptography</a></dt>
 <dd>
 <div class="block">Get the current key held in memory</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../Steg/StegMeister.html#getPrimaryStage--">getPrimaryStage()</a></span> - Static method in class Steg.<a href="../Steg/StegMeister.html" title="class in Steg">StegMeister</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/StegMeister.html#getPrimaryStage--">getPrimaryStage()</a></span> - Static method in class steg.<a href="../Steg/StegMeister.html" title="class in steg">StegMeister</a></dt>
 <dd>&nbsp;</dd>
 </dl>
 <a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">E</a>&nbsp;<a href="index-5.html">G</a>&nbsp;<a href="index-6.html">I</a>&nbsp;<a href="index-7.html">M</a>&nbsp;<a href="index-8.html">O</a>&nbsp;<a href="index-9.html">S</a>&nbsp;<a href="index-10.html">T</a>&nbsp;</div>
