+++ docs/javadoc/index-files/index-7.html
@@ -98,15 +98,15 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </a>
 <h2 class="title">M</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../Steg/StegMeister.html#main-java.lang.String:A-">main(String[])</a></span> - Static method in class Steg.<a href="../Steg/StegMeister.html" title="class in Steg">StegMeister</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/StegMeister.html#main-java.lang.String:A-">main(String[])</a></span> - Static method in class steg.<a href="../Steg/StegMeister.html" title="class in steg">StegMeister</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Cryptography.html#maxKeySize--">maxKeySize()</a></span> - Method in class Steg.Cryptography.<a href="../Steg/Cryptography/Cryptography.html" title="class in Steg.Cryptography">Cryptography</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Cryptography.html#maxKeySize--">maxKeySize()</a></span> - Method in class steg.cryptography.<a href="../Steg/Cryptography/Cryptography.html" title="class in steg.cryptography">Cryptography</a></dt>
 <dd>
 <div class="block">Check the maximum key size supported by the jre</div>
 </dd>
-<dt><a href="../Steg/Steganography/Model.html" title="class in Steg.Steganography"><span class="typeNameLink">Model</span></a> - Class in <a href="../Steg/Steganography/package-summary.html">Steg.Steganography</a></dt>
+<dt><a href="../Steg/Steganography/Model.html" title="class in steg.steganography"><span class="typeNameLink">Model</span></a> - Class in <a href="../Steg/Steganography/package-summary.html">steg.steganography</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../Steg/Steganography/Model.html#Model--">Model()</a></span> - Constructor for class Steg.Steganography.<a href="../Steg/Steganography/Model.html" title="class in Steg.Steganography">Model</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/Steganography/Model.html#Model--">Model()</a></span> - Constructor for class steg.steganography.<a href="../Steg/Steganography/Model.html" title="class in steg.steganography">Model</a></dt>
 <dd>&nbsp;</dd>
 </dl>
 <a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">E</a>&nbsp;<a href="index-5.html">G</a>&nbsp;<a href="index-6.html">I</a>&nbsp;<a href="index-7.html">M</a>&nbsp;<a href="index-8.html">O</a>&nbsp;<a href="index-9.html">S</a>&nbsp;<a href="index-10.html">T</a>&nbsp;</div>
