+++ docs/javadoc/index-files/index-9.html
@@ -98,29 +98,29 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </a>
 <h2 class="title">S</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Cryptography.html#saveKey-java.lang.String-">saveKey(String)</a></span> - Method in class Steg.Cryptography.<a href="../Steg/Cryptography/Cryptography.html" title="class in Steg.Cryptography">Cryptography</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Cryptography.html#saveKey-java.lang.String-">saveKey(String)</a></span> - Method in class steg.cryptography.<a href="../Steg/Cryptography/Cryptography.html" title="class in steg.cryptography">Cryptography</a></dt>
 <dd>
 <div class="block">Save the current key to a file for later use</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Cryptography.html#setKey-javax.crypto.SecretKey-">setKey(SecretKey)</a></span> - Method in class Steg.Cryptography.<a href="../Steg/Cryptography/Cryptography.html" title="class in Steg.Cryptography">Cryptography</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Cryptography.html#setKey-javax.crypto.SecretKey-">setKey(SecretKey)</a></span> - Method in class steg.cryptography.<a href="../Steg/Cryptography/Cryptography.html" title="class in steg.cryptography">Cryptography</a></dt>
 <dd>
 <div class="block">Set the secret key manually</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Encrypt.html#setKeyRand--">setKeyRand()</a></span> - Method in class Steg.Cryptography.<a href="../Steg/Cryptography/Encrypt.html" title="class in Steg.Cryptography">Encrypt</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/Cryptography/Encrypt.html#setKeyRand--">setKeyRand()</a></span> - Method in class steg.cryptography.<a href="../Steg/Cryptography/Encrypt.html" title="class in steg.cryptography">Encrypt</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../Steg/StegMeister.html#start-javafx.stage.Stage-">start(Stage)</a></span> - Method in class Steg.<a href="../Steg/StegMeister.html" title="class in Steg">StegMeister</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/StegMeister.html#start-javafx.stage.Stage-">start(Stage)</a></span> - Method in class steg.<a href="../Steg/StegMeister.html" title="class in steg">StegMeister</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="../Steg/package-summary.html">Steg</a> - package Steg</dt>
+<dt><a href="../Steg/package-summary.html">steg</a> - package steg</dt>
 <dd>&nbsp;</dd>
-<dt><a href="../Steg/Cryptography/package-summary.html">Steg.Cryptography</a> - package Steg.Cryptography</dt>
+<dt><a href="../Steg/Cryptography/package-summary.html">steg.cryptography</a> - package steg.cryptography</dt>
 <dd>&nbsp;</dd>
-<dt><a href="../Steg/Interface/package-summary.html">Steg.Interface</a> - package Steg.Interface</dt>
+<dt><a href="../Steg/Interface/package-summary.html">steg.ui</a> - package steg.ui</dt>
 <dd>&nbsp;</dd>
-<dt><a href="../Steg/Steganography/package-summary.html">Steg.Steganography</a> - package Steg.Steganography</dt>
+<dt><a href="../Steg/Steganography/package-summary.html">steg.steganography</a> - package steg.steganography</dt>
 <dd>&nbsp;</dd>
-<dt><a href="../Steg/StegMeister.html" title="class in Steg"><span class="typeNameLink">StegMeister</span></a> - Class in <a href="../Steg/package-summary.html">Steg</a></dt>
+<dt><a href="../Steg/StegMeister.html" title="class in steg"><span class="typeNameLink">StegMeister</span></a> - Class in <a href="../Steg/package-summary.html">steg</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../Steg/StegMeister.html#StegMeister--">StegMeister()</a></span> - Constructor for class Steg.<a href="../Steg/StegMeister.html" title="class in Steg">StegMeister</a></dt>
+<dt><span class="memberNameLink"><a href="../Steg/StegMeister.html#StegMeister--">StegMeister()</a></span> - Constructor for class steg.<a href="../Steg/StegMeister.html" title="class in steg">StegMeister</a></dt>
 <dd>&nbsp;</dd>
 </dl>
 <a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">E</a>&nbsp;<a href="index-5.html">G</a>&nbsp;<a href="index-6.html">I</a>&nbsp;<a href="index-7.html">M</a>&nbsp;<a href="index-8.html">O</a>&nbsp;<a href="index-9.html">S</a>&nbsp;<a href="index-10.html">T</a>&nbsp;</div>
