+++ docs/javadoc/overview-frame.html
@@ -26,10 +26,10 @@
 <div class="indexContainer">
 <h2 title="Packages">Packages</h2>
 <ul title="Packages">
-<li><a href="Steg/package-frame.html" target="packageFrame">Steg</a></li>
-<li><a href="Steg/Cryptography/package-frame.html" target="packageFrame">Steg.Cryptography</a></li>
-<li><a href="Steg/Interface/package-frame.html" target="packageFrame">Steg.Interface</a></li>
-<li><a href="Steg/Steganography/package-frame.html" target="packageFrame">Steg.Steganography</a></li>
+<li><a href="Steg/package-frame.html" target="packageFrame">steg</a></li>
+<li><a href="Steg/Cryptography/package-frame.html" target="packageFrame">steg.cryptography</a></li>
+<li><a href="Steg/Interface/package-frame.html" target="packageFrame">steg.ui</a></li>
+<li><a href="Steg/Steganography/package-frame.html" target="packageFrame">steg.steganography</a></li>
 </ul>
 </div>
 <p>&nbsp;</p>
