+++ docs/javadoc/overview-summary.html
@@ -104,19 +104,19 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </tr>
 <tbody>
 <tr class="altColor">
-<th class="colFirst" scope="row"><a href="Steg/package-summary.html">Steg</a></th>
+<th class="colFirst" scope="row"><a href="Steg/package-summary.html">steg</a></th>
 <td class="colLast">&nbsp;</td>
 </tr>
 <tr class="rowColor">
-<th class="colFirst" scope="row"><a href="Steg/Cryptography/package-summary.html">Steg.Cryptography</a></th>
+<th class="colFirst" scope="row"><a href="Steg/Cryptography/package-summary.html">steg.cryptography</a></th>
 <td class="colLast">&nbsp;</td>
 </tr>
 <tr class="altColor">
-<th class="colFirst" scope="row"><a href="Steg/Interface/package-summary.html">Steg.Interface</a></th>
+<th class="colFirst" scope="row"><a href="Steg/Interface/package-summary.html">steg.ui</a></th>
 <td class="colLast">&nbsp;</td>
 </tr>
 <tr class="rowColor">
-<th class="colFirst" scope="row"><a href="Steg/Steganography/package-summary.html">Steg.Steganography</a></th>
+<th class="colFirst" scope="row"><a href="Steg/Steganography/package-summary.html">steg.steganography</a></th>
 <td class="colLast">&nbsp;</td>
 </tr>
 </tbody>
