+++ docs/javadoc/overview-tree.html
@@ -97,10 +97,10 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <h1 class="title">Hierarchy For All Packages</h1>
 <span class="packageHierarchyLabel">Package Hierarchies:</span>
 <ul class="horizontal">
-<li><a href="Steg/package-tree.html">Steg</a>, </li>
-<li><a href="Steg/Cryptography/package-tree.html">Steg.Cryptography</a>, </li>
-<li><a href="Steg/Interface/package-tree.html">Steg.Interface</a>, </li>
-<li><a href="Steg/Steganography/package-tree.html">Steg.Steganography</a></li>
+<li><a href="Steg/package-tree.html">steg</a>, </li>
+<li><a href="Steg/Cryptography/package-tree.html">steg.cryptography</a>, </li>
+<li><a href="Steg/Interface/package-tree.html">steg.ui</a>, </li>
+<li><a href="Steg/Steganography/package-tree.html">steg.steganography</a></li>
 </ul>
 </div>
 <div class="contentContainer">
@@ -110,23 +110,23 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <ul>
 <li class="circle">javafx.application.Application
 <ul>
-<li class="circle">Steg.<a href="Steg/StegMeister.html" title="class in Steg"><span class="typeNameLink">StegMeister</span></a>
+<li class="circle">steg.<a href="Steg/StegMeister.html" title="class in steg"><span class="typeNameLink">StegMeister</span></a>
 <ul>
-<li class="circle">Steg.Interface.<a href="Steg/Interface/Controller.html" title="class in Steg.Interface"><span class="typeNameLink">Controller</span></a></li>
+<li class="circle">steg.ui.<a href="Steg/Interface/Controller.html" title="class in steg.ui"><span class="typeNameLink">Controller</span></a></li>
 </ul>
 </li>
 </ul>
 </li>
-<li class="circle">Steg.Steganography.<a href="Steg/Steganography/BasicDecoder.html" title="class in Steg.Steganography"><span class="typeNameLink">BasicDecoder</span></a></li>
-<li class="circle">Steg.Steganography.<a href="Steg/Steganography/BasicEncoder.html" title="class in Steg.Steganography"><span class="typeNameLink">BasicEncoder</span></a></li>
-<li class="circle">Steg.Cryptography.<a href="Steg/Cryptography/Ciph.html" title="class in Steg.Cryptography"><span class="typeNameLink">Ciph</span></a></li>
-<li class="circle">Steg.Cryptography.<a href="Steg/Cryptography/Cryptography.html" title="class in Steg.Cryptography"><span class="typeNameLink">Cryptography</span></a>
+<li class="circle">steg.steganography.<a href="Steg/Steganography/BasicDecoder.html" title="class in steg.steganography"><span class="typeNameLink">BasicDecoder</span></a></li>
+<li class="circle">steg.steganography.<a href="Steg/Steganography/BasicEncoder.html" title="class in steg.steganography"><span class="typeNameLink">BasicEncoder</span></a></li>
+<li class="circle">steg.cryptography.<a href="Steg/Cryptography/Ciph.html" title="class in steg.cryptography"><span class="typeNameLink">Ciph</span></a></li>
+<li class="circle">steg.cryptography.<a href="Steg/Cryptography/Cryptography.html" title="class in steg.cryptography"><span class="typeNameLink">Cryptography</span></a>
 <ul>
-<li class="circle">Steg.Cryptography.<a href="Steg/Cryptography/Decrypt.html" title="class in Steg.Cryptography"><span class="typeNameLink">Decrypt</span></a></li>
-<li class="circle">Steg.Cryptography.<a href="Steg/Cryptography/Encrypt.html" title="class in Steg.Cryptography"><span class="typeNameLink">Encrypt</span></a></li>
+<li class="circle">steg.cryptography.<a href="Steg/Cryptography/Decrypt.html" title="class in steg.cryptography"><span class="typeNameLink">Decrypt</span></a></li>
+<li class="circle">steg.cryptography.<a href="Steg/Cryptography/Encrypt.html" title="class in steg.cryptography"><span class="typeNameLink">Encrypt</span></a></li>
 </ul>
 </li>
-<li class="circle">Steg.Steganography.<a href="Steg/Steganography/Model.html" title="class in Steg.Steganography"><span class="typeNameLink">Model</span></a></li>
+<li class="circle">steg.steganography.<a href="Steg/Steganography/Model.html" title="class in steg.steganography"><span class="typeNameLink">Model</span></a></li>
 </ul>
 </li>
 </ul>
