+++ docs/javadoc/type-search-index.js
@@ -1,1 +1,1 @@
-typeSearchIndex = [{"p":"Steg.Steganography","l":"BasicDecoder"},{"p":"Steg.Steganography","l":"BasicEncoder"},{"p":"Steg.Cryptography","l":"Ciph"},{"p":"Steg.Interface","l":"Controller"},{"p":"Steg.Cryptography","l":"Cryptography"},{"p":"Steg.Cryptography","l":"Decrypt"},{"p":"Steg.Cryptography","l":"Encrypt"},{"p":"Steg.Steganography","l":"Model"},{"p":"Steg","l":"StegMeister"}]
+typeSearchIndex = [{"p":"steg.steganography","l":"BasicDecoder"},{"p":"steg.steganography","l":"BasicEncoder"},{"p":"steg.cryptography","l":"Ciph"},{"p":"steg.ui","l":"Controller"},{"p":"steg.cryptography","l":"cryptography"},{"p":"steg.cryptography","l":"Decrypt"},{"p":"steg.cryptography","l":"Encrypt"},{"p":"steg.steganography","l":"Model"},{"p":"steg","l":"StegMeister"}]
