+++ src/test/resources/sampleimages/White_Cliffs_of_Dover_02.JPG
